requirements.txt
# Google Gemini API
google-generativeai

# Telegram API
pyTelegramBotAPI

# to handle tokens
python-dotenv